make sure nil filters don't blow up in the apiClient.do function
This is needed because comparing an interface value to nil is pointless,
as we don't compare the internal, actual value.

See https://groups.google.com/forum/#!topic/golang-nuts/wnH302gBa4I/discussion
for why checking for wrapped nils is a bad idea in general and violates
the concept of interfaces.